refactor(chess): 컨트롤러와 뷰의 이름과 경로를 변경
- ChessController 를 ConsoleController 로 변경
- InputView와 OutputView를 console 패키지로 이동
