+++ src/main/java/chess/ConsoleApplication.java
@@ -1,10 +1,10 @@
 package chess;
 
-import chess.controller.ChessController;
+import chess.controller.ConsoleController;
 
 public class ConsoleApplication {
     public static void main(String[] args) {
-        ChessController controller = new ChessController();
+        ConsoleController controller = new ConsoleController();
         controller.run();
     }
 }
