+++ src/main/java/chess/controller/ConsoleController.java
@@ -3,12 +3,12 @@ package chess.controller;
 import chess.domain.game.ChessGame;
 import chess.domain.game.score.ScoreResult;
 import chess.domain.position.Position;
-import chess.view.InputView;
-import chess.view.OutputView;
+import chess.view.console.InputView;
+import chess.view.console.OutputView;
 import chess.view.dto.BoardDto;
 import chess.view.dto.ConsoleCommandDto;
 
-public class ChessController {
+public class ConsoleController {
 
     public void run() {
         ChessGame chessGame = new ChessGame();
